pyproject.toml
[tool.poetry]
name = "imdb-clone-django-rest"
version = "0.1.0"
license = "MIT"
readme = "README.md"
homepage = "https://github.com/paulrousset/imdb-clone-django-rest"
repository = "https://github.com/paulrousset/imdb-clone-django-rest"
description = "IMDB API clone using django REST framework"
keywords = ["imdb-clone"]
authors = ["Paul <paulrousset@hotmail.fr>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.2"
djangorestframework = "^3.14.0"
pre-commit = "^2.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
